Added load checking on remote machine. If you have few machines which are using the same server for backups, could be handy.
During early connect to backup destination, check for load is performed.
If it's over the limit, process sleeps for 5 minutes. This will happen 3
times, then script will continue with local run only. So we are not going
to miss backup schedule.

Based on the -i/-I parameter, this snapshots will be transferred to remote
system on next run.